Closes the current plotting device and shows the result in the viewer

To be used in combination with pdf_fit or png_fit when exporting images. It performs exactly the same thing as dev.off() but additionaly shows the resulting graph in the viewer pane provided you're using RStudio.

fit.off()

Details

To view the results of PDF exports, the function pdf_convert from package pdftools is used to convert the PDF files into images -- so you need to have installed pdftools to make it work.

In PDFs, only the first page will be viewed.

See also

Author

Laurent Berge

Examples

# Exportation example
# The functions pdf_fit, png_fit, etc, guarantee the right
#  point size of the texts present in the graph.
# But you must give the exact size the graph will take in your final document.
# => first use the function setFplot_page, default is:
# setFplot_page(page = "us", margins = "normal")
# By default the graph takes 100% of the text width

data(us_pub_econ)

tmpFile = file.path(tempdir(), "DISTR -- institutions.png")

png_fit(tmpFile)
plot_distr(~institution, us_pub_econ)
fit.off()
#> Warning: The function 'fit.off' only works with RStudio's viewer--which wasn't found.

# What's the consequence of increasing the point size of the text?
png_fit(tmpFile, pt = 15)
plot_distr(~institution, us_pub_econ)
fit.off()