Aggregate/conditional graphs and automatic layout using formulas

fplot provides automatic plotting of common graphs (distributions, lines, bar plots and boxplots). The syntax uses formulas, allowing aggregate/conditional/weighted graphs with minimum efforts. The many arguments are automatically adjusted to the data in order to provide the nicest and most meaningful graphs.

Details

The core functions is plot_distr to draw distributions. Two other graphical functions are provided for convenience: plot_lines to represent the (usually temporal) evolution of some variables, and plot_box to easily represent conditional boxplots.

It also integrates tools to easily export graphs: pdf_fit and png_fit. In these functions, instead of providing the size of the graphics, you instead give the point size that the text should have in the final document--because an exported graph usually ends up in a document. You can set the size of your document with the function setFplot_page. If you use the function fit.off to close the connection, you will also see how the export looks like in the Viewer pane.

Author

Maintainer: Laurent Berge laurent.berge@u-bordeaux.fr